+++ frontend/package.json
@@ -22,14 +22,10 @@
     "react-hook-form": "^7.43.1",
     "react-js-pagination": "^3.0.3",
     "react-quill": "^2.0.0",
-    "react-redux": "^8.0.5",
     "react-router": "^6.8.1",
     "react-router-dom": "^6.8.1",
     "react-scripts": "5.0.1",
     "recoil": "^0.7.7",
-    "redux": "^4.2.1",
-    "redux-promise": "^0.6.0",
-    "redux-thunk": "^2.4.2",
     "styled-components": "^5.3.6",
     "styled-reset": "^4.4.5",
     "sweetalert2": "^11.7.12",
