+++ frontend/src/Routes.js
@@ -6,7 +6,7 @@ import Calendar from './pages/Calendar';
 import MyPage from './pages/MyPage';
 import Manage from './pages/Manage';
 import Noticeboard from './pages/Noticeboard';
-import MemberDetail from './pages/MemberDetail';
+import MemberDetail from './components/manage/MemberDetail';
 import Login from './pages/Login';
 import CabinetRent from './pages/CabinetRent';
 import UmbrellaRent from './pages/UmbrellaRent';
@@ -31,10 +31,7 @@ function Router() {
             <PrivateRoute authenticated={isLogin} component={<Noticeboard />} />
           }
         />
-        <Route
-          path="/manage/memberDetail/:userNickname"
-          element={<MemberDetail />}
-        />
+        <Route path="/manage/memberinfo/:user" element={<MemberDetail />} />
         <Route path="/rent/umbrellarent" element={<UmbrellaRent />} />
         <Route path="/rent/cabinetrent" element={<CabinetRent />} />
       </Routes>
