--- frontend/src/_actions/datePickerAction.js
@@ -1,8 +1,0 @@
-import { SELECT_DATE } from './types';
-
-export function selectDate(data) {
-  return {
-    type: SELECT_DATE,
-    payload: data,
-  };
-}
