--- frontend/src/_actions/types.js
@@ -1,7 +1,0 @@
-export const REGISTER_USER = 'REGISTER_USER';
-export const LOGIN_USER = 'LOGIN_USER';
-export const LOGOUT_USER = 'LOGOUT_USER';
-export const INIT_MEMBER = 'INIT_MEMBER';
-export const CHANGE_GRADE = 'CHANGE_GRADE';
-export const DELETE_MEMBER = 'DELETE_MEMBER';
-export const SELECT_DATE = 'SELECT_DATE';
