--- frontend/src/_reducers/UserReducer.js
@@ -1,14 +1,0 @@
-import { LOGIN_USER, LOGOUT_USER, REGISTER_USER } from '../_actions/types';
-
-export default function userReducer(state = {}, action) {
-  switch (action.type) {
-    case REGISTER_USER:
-      return { ...state, registerSuccess: action.payload };
-    case LOGIN_USER:
-      return { ...state, loginSuccess: action.payload };
-    case LOGOUT_USER:
-      return { ...state, logoutSuccess: action.payload };
-    default:
-      return state;
-  }
-}
