--- frontend/src/_reducers/datePickerReducer.js
@@ -1,10 +1,0 @@
-import { SELECT_DATE } from '../_actions/types';
-
-export default function datePickerReducer(state = {}, action) {
-  switch (action.type) {
-    case SELECT_DATE:
-      return { ...state, selectedDay: action.payload };
-    default:
-      return { ...state, selectedDay: new Date() };
-  }
-}
