+++ frontend/src/atom/date.js
@@ -2,5 +2,5 @@ import { atom } from 'recoil';
 
 export const date = atom({
   key: 'date',
-  default: { selectedDate: new Date() },
+  default: new Date(),
 });
