+++ frontend/src/index.js
@@ -6,18 +6,9 @@ import { ThemeProvider } from 'styled-components';
 import reportWebVitals from './reportWebVitals';
 
 import Routes from './Routes';
-import { applyMiddleware, createStore } from 'redux';
-import promiseMiddleware from 'redux-promise';
-import reduxThunk from 'redux-thunk';
-import { Provider } from 'react-redux';
-import rootReducer from './_reducers/RootReducer';
+
 import { worker } from './mocks/browsers';
 import { RecoilRoot } from 'recoil';
-
-const createStoreWidthMiddleware = applyMiddleware(
-  promiseMiddleware,
-  reduxThunk,
-)(createStore);
 
 if (process.env.NODE_ENV === 'development') {
   worker.start();
@@ -29,15 +20,7 @@ root.render(
     <GlobalStyle />
     <ThemeProvider theme={Theme}>
       <RecoilRoot>
-        <Provider
-          store={createStoreWidthMiddleware(
-            rootReducer,
-            window.__REDUX_DEVTOOLS_EXTENSION__ &&
-              window.__REDUX_DEVTOOLS_EXTENSION__(),
-          )}
-        >
-          <Routes />
-        </Provider>
+        <Routes />
       </RecoilRoot>
     </ThemeProvider>
   </>,
